+++ app/templates/server/config/socketio(socketio).js
@@ -29,7 +29,7 @@ module.exports = function (socketio) {
 
   // We can authenticate socket.io users and access their token through socket.handshake.decoded_token
   //
-  // 1. You will need to send the token in `client/components/socket/socket.service.js`
+  // 1. You will need to send the token in `client/sdk/socket/socket.service.js`
   //
   // 2. Require authentication here:
   // socketio.use(require('socketio-jwt').authorize({