+++ app/templates/server/routes.js
@@ -4,7 +4,7 @@
 
 'use strict';
 
-var errors = require('./components/errors');
+var errors = require('./sdk/errors');
 var path = require('path');
 
 module.exports = function(app) {
@@ -16,7 +16,7 @@ module.exports = function(app) {
   app.use('/auth', require('./auth'));
   <% } %>
   // All undefined asset or api routes should return a 404
-  app.route('/:url(api|auth|components|app|bower_components|assets)/*')
+  app.route('/:url(api|auth|sdk|app|bower_components|assets)/*')
    .get(errors[404]);
 
   // All other routes should redirect to the index.html